cron-job changes

+++ backend/server.js
@@ -21,7 +21,7 @@ const deleteUser=async()=>{
     let response=await User.find();
     for (const data of response) {
         let time=Date.now()-data.date.getTime();//7200000
-        if(time>7200000){//2 hours =7200000 milliseconds
+        if(time>7200000 && data.confirmation===false){//2 hours =7200000 milliseconds
             await User.deleteOne({_id:date._id})
         }
     }
